Organize protocol runtime and configs
