+++ configs/protocol/sharc_uv/curing_scan.yaml
