+++ configs/protocol/sharc_uv/motion_scan.yaml
@@ -1,6 +1,6 @@
 # SHARC UV motion-only scan.
 #
-# Exercises the same nested-plate scan path as sharc_uv_curing_scan.yaml but
+# Exercises the same nested-plate scan path as curing_scan.yaml but
 # uses uv_curing.health_check instead of cure. With configs/gantry/cub_sharc.yaml
 # keeping uv_curing offline, this moves the gantry over all 96 wells without
 # sending OmniCure cure commands or opening UV exposure.
