+++ configs/protocol/sterling/vial_scan.yaml
