+++ docs/reference/src/protocol_engine/index.md
@@ -11,6 +11,7 @@
 - [src.protocol_engine.loader](loader.md)
 - [src.protocol_engine.measurements](measurements.md)
 - [src.protocol_engine.protocol](protocol.md)
+- [src.protocol_engine.runtime](runtime.md)
 - [src.protocol_engine.scan_args](scan_args.md)
 - [src.protocol_engine.registry](registry.md)
 - [src.protocol_engine.setup](setup.md)
