+++ setup/validate_setup.py
@@ -7,7 +7,7 @@ Example:
     python setup/validate_setup.py \
         ../BU-Configs/configs/gantry/cub_xl_asmi.yaml \
         ../BU-Configs/configs/deck/asmi_deck.yaml \
-        ../BU-Configs/configs/protocol/asmi_move_a1.yaml
+        ../BU-Configs/configs/protocol/asmi/move_a1.yaml
 """
 
 from __future__ import annotations
@@ -33,7 +33,7 @@ def main() -> None:
         print("  python setup/validate_setup.py \\")
         print("    ../BU-Configs/configs/gantry/cub_xl_asmi.yaml \\")
         print("    ../BU-Configs/configs/deck/asmi_deck.yaml \\")
-        print("    ../BU-Configs/configs/protocol/asmi_move_a1.yaml")
+        print("    ../BU-Configs/configs/protocol/asmi/move_a1.yaml")
         sys.exit(1)
 
     gantry_path, deck_path, protocol_path = sys.argv[1:4]
