+++ src/protocol_engine/__init__.py
@@ -1,7 +1,8 @@
 from protocol_engine.errors import ProtocolExecutionError, ProtocolLoaderError
 from protocol_engine.loader import load_protocol_from_yaml, load_protocol_from_yaml_safe
-from protocol_engine.protocol import Protocol, ProtocolContext, ProtocolStep
+from protocol_engine.protocol import Protocol
 from protocol_engine.registry import CommandRegistry, protocol_command
+from protocol_engine.runtime import ProtocolContext, ProtocolStep
 
 __all__ = [
     "CommandRegistry",
